portfolio-backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.1
flask-jwt-extended==4.6.0
pymongo==4.10.1
python-dotenv==1.0.1
dnspython==2.7.0
ua-parser==0.18.0
requests==2.32.3
bcrypt==4.2.0
mangum==0.17.0
boto3==1.34.0
